refactor(dashboard): extract Mastio auth sub-router (F-B-201 PR-2/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-201.md

Sprint F-B-201 PR-2 of 10. base=main. Mirrors Court PR-2 (#841).

Extracts 5 password-auth routes from mcp_proxy/dashboard/router.py
into a dedicated sub-router. Pure cut+paste; behaviour-equivalent.

  mcp_proxy/dashboard/auth_routes.py (new, 268 LOC, 5 routes):
    GET  /proxy/login      login form + redirect-to-register if pristine
    POST /proxy/login      H9 lockout + rate-limit + bcrypt + audit
    POST /proxy/logout     CSRF-gated session clear (F-B-9)
    GET  /proxy/register   one-shot admin password form
    POST /proxy/register   set_admin_password + audit + redirect-to-login

  mcp_proxy/dashboard/_helpers.py (+20 LOC, 2 new helpers):
    _post_login_redirect    (was inline at router.py:191 since #54)
    _load_display_name      (was inline at router.py:5037 since #190)
    Both are pure async helpers consumed by auth_routes + remaining
    inline OIDC handler in router.py.

  mcp_proxy/dashboard/router.py (5052 -> 4828 LOC, -224 net):
    - 5 route bodies removed; breadcrumb at the old positions.
    - 2 helper bodies moved to _helpers.py; breadcrumb.
    - One new include_router(_auth_routes.router) call.

Behavioural delta: zero. H9 audit lockout + rate limit gate
preserved on POST /login. F-B-9 CSRF gate preserved on POST /logout.
F-A-507 insecure-default password gate (registered as pre-existing
on set_admin_password) unaffected.

Scope note: OIDC routes (/proxy/oidc/start, /proxy/oidc/callback) stay
inline in router.py for now. Sister-file split with mcp_proxy/dashboard/
oidc.py is already partial (helpers there, handler still in router.py);
a future PR can decide whether to consolidate.

Tests:
  - pytest tests/test_h9_login_lockout.py tests/test_fb9_proxy_logout_csrf.py
    tests/test_proxy_local_password_toggle.py
    tests/test_proxy_dashboard_oidc.py
    tests/test_proxy_dashboard_display_name.py
    -n auto --dist=loadfile -q
  - 52 passed.

Router LOC trajectory:
  Pre-Sprint:    66 routes, 5106 LOC
  Post-PR-1 #859: 66 routes, 5052 LOC
  Post-PR-2 (this): 61 routes, 4828 LOC
  Target post-PR-10: ~3 routes, ~120 LOC aggregator.

Next PR-3: /proxy/setup (broker uplink wizard, ~530 LOC, 3 routes
across /setup GET + POST + /setup/test-connection) into setup_routes.py.
